+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
-# 1.2.1 - In progress
+# 1.2.2 - In progress
+
+- [Feature] Add ability edit uid and zero block
+- [Feature] Add support adaptive monochrome icon
+- [Feature] Add save as on nfc editor
+- [Feature] Add dialog on exit from nfc editor
+- [REFACTOR] Migrate synchronization task graph to TaskSynchronizationComponent
+- [REFACTOR] Updater: new request model
+- [REFACTOR] Redesign card when we waiting update
+- [FIX] Rename updater folder
+- [FIX] Choose app theme from selected theme, not is light
+- [FIX] Check total bytes from server(less 0 mean error from server)
+
+# 1.2.1
 
 - [REFACTOR] Change Flipper logo
 - [REFACTOR] Redesign Full Device Information
@@ -30,6 +43,8 @@
 - [FIX] Subghz provisioning while geoip is null
 - [FIX] Separate screen for edit key
 - [FIX] Synchronization with shadow files
+- [FIX] Migrate wearos model to protobuf
+- [FIX] Double synchronization
 
 # 1.2.0